requirements.txt
pandas==1.3.5
redis==4.3.4
numpy==1.21.6
pycryptodome==3.15.0
pytz==2022.4
tqdm==4.64.1
plotly==5.10.0
py-cpuinfo==8.0.0
pandas-ta==0.3.14b0
scipy==1.7.3
requests~=2.28.1
setuptools==65.4.1
python-dotenv==0.21.0
